webapi/webapi-rawsockets-w3c-tests/rawsockets/UDPSocket_attribute.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Shentu.Jiazhen <jiazhenx.shentu@intel.com>

-->

<meta charset='utf-8'>
<title>Raw Socket Test: UDPSocket attribute</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/raw-sockets/#attributes">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  assert_true(socket.addressReuse, "UDPSocket.addressReuse default is");
}, "Check if the default value of UDPSocket.addressReuse is true");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  assert_equals(socket.bufferedAmount, 0, "The value of socket.bufferedAmount");
}, "Check if the value of UDPSocket.bufferedAmount is the number of bytes which have previously been buffered by calls to the send methods of this socket");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var config = {
    "localAddress": "127.0.0.1",
    "remoteAddress": "239.255.255.250"
  };
  var socket = new namespace.UDPSocket(config);
  assert_equals(socket.localAddress , "127.0.0.1", "UDPSocket.localAddress");
}, "Check if UDPSocket.localAddress can be set by the constructor's options argument's valid localAddress member");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var config = {
    "localAddress": "test",
    "remoteAddress": "239.255.255.250"
  };
  assert_throws(
    "InvalidAccessError",
    function () {var socket = new namespace.UDPSocket(config);},
    "InvalidAccessError DOM exception."
  );
}, "Check if UDPSocket throw a DOM exception invoked with the constructor's options argument's invalid localAddress member");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var config = {
    "remoteAddress": "239.255.255.250"
  };
  var socket = new namespace.UDPSocket(config);
  assert_not_equals(socket.localAddress, null, "UDPSocket.localAddress");
}, "Check if UDPSocket.localAddress is not null when the remoteAddress member is present");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  assert_equals(socket.localAddress, null, "UDPSocket.localAddress");
}, "Check if UDPSocket.localAddress is null without the localAddress and the remoteAddress members");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  assert_false(socket.loopback, "UDPSocket.loopback");
}, "Check if the default value of UDPSocket.loopback is false");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var config = {"remoteAddress": "239.255.255.250"};
  var socket = new namespace.UDPSocket(config);
  assert_equals(socket.remoteAddress, "239.255.255.250", "UDPSocket.remoteAddress");
}, "Check if UDPSocket.remoteAddress can be set by the constructor's options argument's valid remoteAddress member");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var config = {"remoteAddress": "test"};
  assert_throws(
    "InvalidAccessError",
    function () {var socket = new namespace.UDPSocket(config);},
    "InvalidAccessError DOM exception."
  );
}, "Check if UDPSocket throw a DOM exception invoked with the constructor's options argument's invalid remoteAddress member");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  assert_equals(socket.remoteAddress, null, "UDPSocket.remoteAddress");
}, "Check if UDPSocket.remoteAddress is null when not stated by the options argument of the constructor");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var config = {
    "remoteAddress": "239.255.255.250",
    "remotePort": 80
  };
  var socket = new namespace.UDPSocket(config);
  assert_equals(socket.remotePort, 80, "UDPSocket.remotePort");
}, "Check if UDPSocket.remotePort can be set by the constructor's options argument's valid remotePort member");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var config = {
    "remoteAddress": "239.255.255.250",
    "remotePort": "test"
  };
  assert_throws(
    "InvalidAccessError",
    function(){var socket = new namespace.UDPSocket(config);},
    "InvalidAccessError DOM exception."
  );
}, "Check if UDPSocket throw a DOM exception invoked with the constructor's options argument's invalid remotePort member");

test(function () {
  assert_true(!!namespace, "UDPSocket interface support");
  var socket = new namespace.UDPSocket();
  assert_equals(socket.remotePort, null, "UDPSocket.remotePort");
}, "Check if UDPSocket.remotePort is null when not stated by the options argument of the constructor");

</script>
